requirements.txt
-i https://pypi.org/simple
cachetools==5.3.2 ; python_version >= '3.7'
certifi==2023.11.17 ; python_version >= '3.6'
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
google-api-core[grpc]==2.15.0 ; python_version >= '3.7'
google-auth==2.27.0 ; python_version >= '3.7'
google-cloud-pubsub==2.19.0
googleapis-common-protos==1.62.0 ; python_version >= '3.7'
grpc-google-iam-v1==0.13.0 ; python_version >= '3.7'
grpcio==1.60.0 ; python_version >= '3.7'
grpcio-status==1.60.0 ; python_version >= '3.6'
idna==3.6 ; python_version >= '3.5'
proto-plus==1.23.0 ; python_version >= '3.6'
protobuf==4.25.2 ; python_version >= '3.8'
pyasn1==0.5.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
requests==2.31.0 ; python_version >= '3.7'
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
slack-bolt==1.18.1
slack-sdk==3.26.2 ; python_version >= '3.6'
urllib3==2.1.0 ; python_version >= '3.8'
